Replace AssessmentCategory model with category_name field
Note: category_id is still in the schema, but should get dropped by the migration, so we made need a commit to update that.

+++ db/migrate/20150203151605_cleanup_old_fields.rb
@@ -22,9 +22,17 @@ class CleanupOldFields < ActiveRecord::Migration
     remove_column :courses, :late_penalty_old, :float
     remove_column :courses, :version_penalty_old, :float
 
-    remove_column :submissions, :tweak_old, :float
-    remove_column :submissions, :ignored_old, :boolean
-    remove_column :submissions, :absolute_tweak, :boolean
+    if column_exists? :submissions, :tweak_old
+        remove_column :submissions, :tweak_old, :float
+    end
+    
+    if column_exists? :submissions, :ignored_old
+        remove_column :submissions, :ignored_old, :boolean
+    end
+
+    if column_exists? :submissions, :absolute_tweak  
+        remove_column :submissions, :absolute_tweak, :boolean
+    end
 
   end
 end
